Move files & fix CS & Stan issues
